change the xrefmissing.py to output upstream commits
Change the xrefmissing.py to output upstream commit hexshas in the table
print to the stdout. Update the blacklist feature so that the blacklist
entries are upstream hexshas. Include the columns 'Missing_commit' and
'Based_on_commit' to the csv output. Do not output (an empty) csv file
if there were no missing commits. Update the documentation.